requirements.txt
Flask-MySQLdb
Flask
flask-restful
PyYAML
requests
marshmallow
